LF-4671 moved weatherApi to /store/api folder to grpup all API's together
